GitHub Fix 2.0
Reorganizes the files within the main folder (2022f-team6) to include all project files and previous assets. When opening project in Unity, be sure to select "The Rock and The Comet" folder, since that is the official project folder now, whereas "2022f-team6" should now properly just be the repo's folder/name.