Reworking Audio Player
I've begun redoing the audio player to make it easier to use/cleaner code. All sounds will now be given to a audio manager that will be called to play them from other scripts. The audio manager uses a custom sound class as well to store the information we'll want for sounds (can add more as needed)